fix a couple compile errors + add logic to check the exception level

+++ bootloader/src/aarch64/registers.rs
@@ -1,4 +1,6 @@
-#[derive(Debug)]
+use core::arch::asm;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ExceptionLevel {
     EL0,
     EL1,
@@ -12,7 +14,7 @@ pub fn current_el() -> ExceptionLevel {
     unsafe {
         asm!("mrs {:x}, CurrentEL", out(reg) el, options(nomem, nostack));
     }
-    match ((el >> 2) & 0b11) as u8 {
+    match ((el >> 2) & 0b11) {
         0 => ExceptionLevel::EL0,
         1 => ExceptionLevel::EL1,
         2 => ExceptionLevel::EL2,
